refactor: models name changed (group, message)
