+++ src/service/user/group/leaveGroup/leaveGroup.service.ts
@@ -1,5 +1,5 @@
 import userCredentials from "../../../../models/userCredentialsModel";
-import Group from "../../../../models/group";
+import Group from "../../../../models/groupModel";
 import { RESPONSE_CODE } from "../../../utils/constants";
 
 export async function leaveGroupService(req: any, res: any, next: any) {
